# Add C code to RoundHalfToEven op.