@notoraptor 2017-01-24
Updated. RoundHalfToEven is now hardcoded and no longer depends on NumPy functions. I have removed one "if" block, compared to the NumPy implementation: https://github.com/numpy/numpy/blob/c90d7c94fd2077d0beca48fa89a423da2b0bb663/numpy/core/src/npymath/npy_math.c.src#L275