@nouiz 2017-01-24
I think the old travis error can be caused by the compilation on an older numpy, that didn't provided that function. But I think adding the code directly is good.