@notoraptor 2017-01-24
Tested on GPU with success (`theano/tensor/tests/test_basic.py:RoundHalfToEvenTester` with float32 and float64):

```
(python2) ~/mila/dev/git/theano$ theano-cache purge
(python2) ~/mila/dev/git/theano$ CUDA_LAUNCH_BLOCKING=1 THEANO_FLAGS=device=cuda,floatX=float32 nosetests -vs theano/tensor/tests/test_basic.py:RoundHalfToEvenTester
Using cuDNN version 5105 on context None
Mapped name None to device cuda: GeForce 840M (0000:0A:00.0)
test_bad_build (theano.tensor.tests.test_basic.RoundHalfToEvenTester) ... ok
test_bad_runtime (theano.tensor.tests.test_basic.RoundHalfToEvenTester) ... ok
test_good (theano.tensor.tests.test_basic.RoundHalfToEvenTester) ... ok
test_grad (theano.tensor.tests.test_basic.RoundHalfToEvenTester) ... ok
test_grad_none (theano.tensor.tests.test_basic.RoundHalfToEvenTester) ... ok

----------------------------------------------------------------------
Ran 5 tests in 26.575s

OK
(python2) ~/mila/dev/git/theano$ CUDA_LAUNCH_BLOCKING=1 THEANO_FLAGS=device=cuda,floatX=float64 nosetests -vs theano/tensor/tests/test_basic.py:RoundHalfToEvenTester
Using cuDNN version 5105 on context None
Mapped name None to device cuda: GeForce 840M (0000:0A:00.0)
test_bad_build (theano.tensor.tests.test_basic.RoundHalfToEvenTester) ... ok
test_bad_runtime (theano.tensor.tests.test_basic.RoundHalfToEvenTester) ... ok
test_good (theano.tensor.tests.test_basic.RoundHalfToEvenTester) ... ok
test_grad (theano.tensor.tests.test_basic.RoundHalfToEvenTester) ... ok
test_grad_none (theano.tensor.tests.test_basic.RoundHalfToEvenTester) ... ok

----------------------------------------------------------------------
Ran 5 tests in 22.605s

OK
```