@pnavarre 2017-02-06
Please check https://github.com/Theano/Theano/issues/5484. There's seems to be an issue with Python 3 and GPU. I see the test above only used Python 2.
Thanks!